feat: Add keyboard support
 - Add a new file keyboard-support.js
 - Include new functionalities, covers Enter, Escape and
 Backspace keystrokes
 - Adjust index.html to cover the new file

footer: final solution employs a hard to track solution
It would be a lot easier to perform using global variables instead of
collecting the expression from the Screen. However, this approach was
much more valuable as I had to learn how to deal with Regex in JS.
(I personally enjoy regex).